Add automatic updates
This change adds automatic updating of the menu (dependent upon a cron job to deploy the updated site files). It also adds an event to the mysql server to delete the existing menu once daily. The cron job should trigger the updater after the event fires for the day to avoid conflicts